pyproject.toml
[project]
name = "ragandwheather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-qdrant>=1.1.0",
    "langgraph>=1.0.5",
    "langsmith>=0.4.59",
    "pypdf>=6.4.2",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "streamlit>=1.52.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
